+++ src/imio/history/__init__.py
@@ -8,4 +8,3 @@ allow_module('imio.history.utils')
 
 def initialize(context):
     """Initializer called when used as a Zope 2 product."""
-